requirements.txt
bs4                0.0.1
pandas             2.1.4
requests           2.31.0
urllib3            2.1.0
pillow             10.2.0
selenium           4.16.0